% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtfismyip.R
\name{wtfismyip}
\alias{wtfismyip}
\title{wtfismyip}
\usage{
wtfismyip(verbose = FALSE, polite = FALSE)
}
\arguments{
\item{verbose}{should the function be verbose?}

\item{polite}{should the function be polite?}
}
\value{
a data frame or a character vector with info about your internet
}
\description{
What is my IP?
}
\examples{
\dontrun{
wtfismyip()
}

}
